enhance class
