+++ db-tool/src/test/java/com/lizp/transform/TransformEntityExcelTest.java
@@ -1,7 +1,11 @@
 package com.lizp.transform;
 
+import com.alibaba.excel.annotation.ExcelProperty;
+import com.lizp.annotation.TableField;
+import com.lizp.annotation.TableName;
 import com.lizp.client.TransformEntityExcelClient;
 import com.lizp.config.TransformEntityConfig;
+import lombok.Data;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -27,6 +31,32 @@ class TransformEntityExcelTest {
     }
 
 
+    /**
+     * @author lizp4
+     * @date 2022-10-8 10:14
+     */
+    @Data
+    public static class Origin {
 
+        @ExcelProperty(value = "name")
+        private String name;
 
+        @ExcelProperty(value = "is_good")
+        private Integer good;
+    }
+
+    /**
+     * @author lizp4
+     * @date 2022-10-8 10:14
+     */
+    @Data
+    @TableName("test_output")
+    public static class Output {
+
+        @TableField("output_name")
+        private String outputName;
+
+        @TableField("output_good")
+        private Integer outputGood;
+    }
 }
